+++ R/weightit.R
@@ -11,7 +11,9 @@ weightit <- function(formula, data, method, estimand = "ATE", stabilize = FALSE,
       stop("formula must be a formula relating treatment to covariates.", call. = FALSE)
     }
     if (missing(data)) {
-      stop("Data must be specified.", call. = FALSE)}
+      data <- environment(formula)
+      #stop("Data must be specified.", call. = FALSE)
+    }
     if (!is.data.frame(data)) {
       stop("Data must be a data.frame.", call. = FALSE)}
   }
@@ -26,7 +28,6 @@ weightit <- function(formula, data, method, estimand = "ATE", stabilize = FALSE,
       else stop("The name supplied to ps is not the name of a variable in data.", call. = FALSE)
     }
   }
-
 
   ##Process method
   bad.method <- FALSE
@@ -48,25 +49,31 @@ weightit <- function(formula, data, method, estimand = "ATE", stabilize = FALSE,
   #Process treat and covs from formula and data
   tt <- terms(formula)
   attr(tt, "intercept") <- 0
-  if (is.na(match(rownames(attr(tt, "factors"))[1], names(data)))) {
-    stop(paste0("The given response variable, \"", rownames(attr(tt, "factors"))[1], "\", is not a variable in data."))
+  if (is.na(match(all.vars(tt[[2]]), names(data)))) {
+    stop(paste0("The given response variable, \"", all.vars(tt[[2]]), "\", is not a variable in data."))
   }
-  vars.mentioned <- unlist(lapply(attr(tt, "variables")[-1], function(x) if (length(x) > 1) paste(x[-1]) else paste(x)))
-  m.try <- try({mf <- model.frame(tt, data)}, TRUE)
-  if (class(m.try) == "try-error") {
-    stop(paste0(c("All variables of formula must be variables in data.\nVariables not in data: ",
-                  paste(attr(tt, "term.labels")[is.na(match(attr(tt, "term.labels"), names(data)))], collapse=", "))), call. = FALSE)}
+  vars.mentioned <- all.vars(tt)
+  tryCatch({mf <- model.frame(tt, data, na.action = na.pass)}, error = function(e) {
+    stop(paste0(c("All variables in formula must be variables in data.\nVariables not in data: ",
+                  paste(vars.mentioned[is.na(match(vars.mentioned, names(data)))], collapse=", "))), call. = FALSE)})
+
   treat <- model.response(mf)
-  covs <- data[!is.na(match(names(data), vars.mentioned[vars.mentioned != rownames(attr(tt, "factors"))[1]]))]
+  covs <- data[!is.na(match(names(data), vars.mentioned[vars.mentioned != all.vars(tt[[2]])]))]
 
   n <- nrow(data)
 
+  if (any(is.na(covs)) || nrow(covs) != n) {
+    stop("No missing values are allowed in the covariates.", call. = FALSE)
+  }
+  if (any(is.na(treat)) || length(treat) != n) {
+    stop("No missing values are allowed in the treatment variable.", call. = FALSE)
+  }
   nunique.treat <- nunique(treat)
   if (nunique.treat == 2) {
     treat.type = "binary"
   }
   else if (nunique.treat < 2) {
-    stop("treatment must have at least two unique values.", call. = FALSE)
+    stop("The treatment must have at least two unique values.", call. = FALSE)
   }
   else if (is.factor(treat) || is.character(treat)) {
     treat.type = "multinomial"
@@ -98,7 +105,7 @@ weightit <- function(formula, data, method, estimand = "ATE", stabilize = FALSE,
   acceptable.exacts <- names(data)
   exact.vars <- character(0)
 
-  if (missing(exact)) exact.factor <- factor(rep(1, n))
+  if (missing(exact) || length(exact) == 0) exact.factor <- factor(rep(1, n))
   else if (!is.atomic(exact)) bad.exact <- TRUE
   else if (is.character(exact) && all(exact %in% acceptable.exacts)) {
     exact.factor <- factor(apply(data[exact], 1, paste, collapse = "|"))
@@ -120,6 +127,8 @@ weightit <- function(formula, data, method, estimand = "ATE", stabilize = FALSE,
   if (length(exact.vars) > 0) {
     formula <- update.formula(formula, as.formula(paste("~ . -", paste(exact.vars, collapse = " - "))))
   }
+
+  call <- match.call()
 
   ## Running models ----
   w <- p.score <- rep(NA_real_, n)
@@ -148,6 +157,7 @@ weightit <- function(formula, data, method, estimand = "ATE", stabilize = FALSE,
         }
         else if (treat.type == "multinomial") {
           estimand <- process.estimand(estimand, c("ATT", "ATE", "ATO"), method, treat.type)
+          process.focal(focal, estimand, treat)
           obj <- weightit2ps.multi(formula = formula,
                                    data = data,
                                    s.weights = s.weights,
@@ -183,6 +193,7 @@ weightit <- function(formula, data, method, estimand = "ATE", stabilize = FALSE,
         }
         else if (treat.type == "multinomial") {
           estimand <- process.estimand(estimand, c("ATT", "ATE"), method, treat.type)
+          process.focal(focal, estimand, treat)
           obj <- weightit2gbm.multi(formula = formula,
                                     data = data,
                                     s.weights = s.weights,
@@ -209,6 +220,7 @@ weightit <- function(formula, data, method, estimand = "ATE", stabilize = FALSE,
         }
         else if (treat.type == "multinomial") {
           estimand <- process.estimand(estimand, c("ATE"), method, treat.type)
+          process.focal(focal, estimand, treat)
           obj <- weightit2cbps.multi(formula = formula,
                                      data = data,
                                      subset = exact.factor == i,
@@ -221,7 +233,7 @@ weightit <- function(formula, data, method, estimand = "ATE", stabilize = FALSE,
                                     data = data,
                                     subset = exact.factor == i,
                                     s.weights = s.weights,
-                                    stabilize = stabilize,
+                                    #stabilize = stabilize,
                                     ...)
 
         }
@@ -239,6 +251,7 @@ weightit <- function(formula, data, method, estimand = "ATE", stabilize = FALSE,
         }
         else if (treat.type == "multinomial") {
           estimand <- process.estimand(estimand, c("ATE"), method, treat.type)
+          process.focal(focal, estimand, treat)
           obj <- weightit2npcbps.multi(formula = formula,
                                        data = data,
                                        subset = exact.factor == i,
@@ -265,6 +278,7 @@ weightit <- function(formula, data, method, estimand = "ATE", stabilize = FALSE,
         }
         else if (treat.type == "multinomial") {
           estimand <- process.estimand(estimand, c("ATT", "ATE"), method, treat.type)
+          process.focal(focal, estimand, treat)
           obj <- weightit2ebal.multi(formula = formula,
                                      data = data,
                                      s.weights = s.weights,
@@ -289,6 +303,7 @@ weightit <- function(formula, data, method, estimand = "ATE", stabilize = FALSE,
         }
         else if (treat.type == "multinomial") {
           estimand <- process.estimand(estimand, c("ATT", "ATE"), method, treat.type)
+          process.focal(focal, estimand, treat)
           obj <- weightit2sbw.multi(formula = formula,
                               data = data,
                               s.weights = s.weights,
@@ -313,6 +328,7 @@ weightit <- function(formula, data, method, estimand = "ATE", stabilize = FALSE,
         }
         else if (treat.type == "multinomial") {
           estimand <- process.estimand(estimand, c("ATE"), method, treat.type)
+          process.focal(focal, estimand, treat)
           obj <- weightit2ebcw.multi(formula = formula,
                                data = data,
                                #s.weights = s.weights,
@@ -372,13 +388,14 @@ weightit <- function(formula, data, method, estimand = "ATE", stabilize = FALSE,
               treat = treat,
               covs = covs,
               data = data,
-              estimand = estimand,
+              estimand = if (treat.type == "continuous") NULL else estimand,
               method = method,
               ps = if (all(is.na(p.score))) NULL else p.score,
               s.weights = s.weights,
               #discarded = NULL,
               treat.type = treat.type,
-              focal = focal)
+              focal = focal,
+              call = call)
   class(out) <- "weightit"
 
   return(out)
@@ -391,8 +408,8 @@ print.weightit <- function(x, ...) {
   cat(paste0(" - number of obs.: ", length(x$weights), "\n"))
   cat(paste0(" - sampling weights: ", ifelse(max(x$s.weights) - min(x$s.weights) < sqrt(.Machine$double.eps),
                                              "none", "present"), "\n"))
-  cat(paste0(" - estimand: ", x$estimand, "\n"))
   cat(paste0(" - treatment: ", ifelse(x$treat.type == "continuous", "continuous", paste0(nunique(x$treat), "-category", ifelse(x$treat.type == "multinomial", paste0(" (", paste(levels(x$treat), collapse = ", "), ")"), ""))), "\n"))
+  if (length(x$estimand) > 0) cat(paste0(" - estimand: ", x$estimand, ifelse(length(x$focal)>0, paste0(" (focal: ", x$focal, ")"), ""), "\n"))
   cat(paste0(" - covariates: ", ifelse(length(names(x$covs)) > 60, "too many to name", paste(names(x$covs), collapse = ", ")), "\n"))
   invisible(x)
 }
@@ -484,7 +501,7 @@ summary.weightit <- function(object, top = 5, ignore.s.weights = FALSE, ...) {
 print.summary.weightit <- function(x, ...) {
   cat("Summary of weights:\n\n")
   cat("- Weight ranges:\n")
-  print.data.frame(round_df(text.box.plot(x$weight.range, 28), 4))
+  print.data.frame(round_df_char(text.box.plot(x$weight.range, 28), 4), ...)
 
   df <- setNames(data.frame(do.call("c", lapply(names(x$weight.top), function(x) c(" ", x))),
                    matrix(do.call("c", lapply(x$weight.top, function(x) c(names(x), round(x, 4)))),
@@ -493,10 +510,10 @@ print.summary.weightit <- function(x, ...) {
   cat(paste("\n- Units with", length(x$weight.top[[1]]), "greatest weights by group:\n"))
   print.data.frame(df, row.names = FALSE)
   cat("\n")
-  print.data.frame(as.data.frame(round(matrix(c(x$weight.ratio, x$coef.of.var), ncol = 2,
+  print.data.frame(round_df_char(as.data.frame(matrix(c(x$weight.ratio, x$coef.of.var), ncol = 2,
                                         dimnames = list(names(x$weight.ratio),
-                                                        c("Ratio", "Coef of Var"))), 4)))
+                                                        c("Ratio", "Coef of Var")))), 4))
   cat("\n- Effective Sample Sizes:\n")
-  print.data.frame(round(x$effective.sample.size, 3))
+  print.data.frame(round_df_char(x$effective.sample.size, 3))
   invisible(x)
 }