+++ NEWS.md
@@ -1,5 +1,11 @@
 WeightIt News and Updates
 ======
+Version 0.3.0
+
+* Added `weightitMSM()` function (and supporting `print()` and `summary()` functions) to estimate weights for marginal structural models with time-varying treatments and covariates.
+
+* Fixed some bugs, including when using CBPS with continuous treatments, and when using `focal` incorrectly.
+
 Version 0.2.0
 
 * Added `method = "sbw"` for stabilized balancing weights